requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
python-dotenv
PyMuPDF
faiss-cpu
tiktoken
jinja2
pdfminer.six
